api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
qdrant-client==1.7.0
sentence-transformers==2.7.0
pydantic==2.5.0
huggingface-hub>=0.15.0
python-multipart==0.0.6
PyJWT==2.8.0
bcrypt==4.1.2
cryptography>=3.4.8
email-validator==2.1.0
asyncpg==0.29.0
redis==5.0.1
httpx==0.25.2
psutil>=5.8.0
prometheus-client==0.19.0
